refactor lambdas
