+++ backend/src/product-service/create-product.ts
@@ -3,7 +3,7 @@ import {
   TransactWriteItemsCommand,
 } from "@aws-sdk/client-dynamodb";
 import { randomUUID } from "crypto";
-import { httpResponse } from "./common/http-responses";
+import { httpResponse } from "../common/http-responses";
 
 const dynamoDbClient = new DynamoDBClient({});
 const PRODUCTS_TABLE_NAME = process.env.PRODUCTS_TABLE_NAME;
